+++ bin/proving-service/src/commands/worker.rs
@@ -1,10 +1,17 @@
 use clap::Parser;
-use tokio::net::TcpListener;
-use tokio_stream::wrappers::TcpListenerStream;
-use tonic_health::server::health_reporter;
-use tracing::{info, instrument};
+use tracing::instrument;
 
-use crate::{api::RpcListener, generated::api_server::ApiServer, utils::MIDEN_PROVING_SERVICE};
+use crate::{
+    utils::MIDEN_PROVING_SERVICE,
+    worker::{batch_prover_worker, tx_prover_worker},
+};
+
+/// Prover type.
+#[derive(Debug, clap::ValueEnum, Clone)]
+enum ProverType {
+    Tx,
+    Batch,
+}
 
 /// Starts a worker.
 #[derive(Debug, Parser)]
@@ -15,11 +22,17 @@ pub struct StartWorker {
     /// The port of the worker
     #[clap(short, long, default_value = "50051")]
     port: u16,
+    /// Prover type
+    ///
+    /// The prover type can be either `tx` or `batch`.
+    #[clap(short, long, default_value = "tx")]
+    prover_type: ProverType,
 }
 
 impl StartWorker {
     /// Starts a worker.
     ///
+    /// It uses the prover_type argument to determine which worker to start.
     /// This method receives the host and port from the CLI and starts a worker on that address.
     /// In case that one of the parameters is not provided, it will default to `0.0.0.0` for the
     /// host and `50051` for the port.
@@ -30,27 +43,15 @@ impl StartWorker {
     #[instrument(target = MIDEN_PROVING_SERVICE, name = "worker:execute")]
     pub async fn execute(&self) -> Result<(), String> {
         let worker_addr = format!("{}:{}", self.host, self.port);
-        let rpc =
-            RpcListener::new(TcpListener::bind(&worker_addr).await.map_err(|err| err.to_string())?);
 
-        info!(
-            "Server listening on {}",
-            rpc.listener.local_addr().map_err(|err| err.to_string())?
-        );
-
-        // Create a health reporter
-        let (mut health_reporter, health_service) = health_reporter();
-
-        // Mark the service as serving
-        health_reporter.set_serving::<ApiServer<RpcListener>>().await;
-
-        tonic::transport::Server::builder()
-            .accept_http1(true)
-            .add_service(tonic_web::enable(rpc.api_service))
-            .add_service(health_service)
-            .serve_with_incoming(TcpListenerStream::new(rpc.listener))
-            .await
-            .map_err(|err| err.to_string())?;
+        match self.prover_type {
+            ProverType::Tx => {
+                tx_prover_worker::start(worker_addr).await?;
+            },
+            ProverType::Batch => {
+                batch_prover_worker::start(worker_addr).await?;
+            },
+        }
 
         Ok(())
     }