+++ bin/proving-service/src/worker/mod.rs
@@ -1,0 +1,2 @@
+pub mod batch_prover_worker;
+pub mod tx_prover_worker;