+++ crates/miden-proving-service-client/src/lib.rs
@@ -15,8 +15,8 @@ use thiserror::Error;
 #[cfg(feature = "tx-prover")]
 pub mod tx_prover;
 
-/// Protobuf definition for the Miden proving service
-pub const TX_PROVER_PROTO: &str = include_str!("../proto/tx_prover.proto");
+#[cfg(feature = "batch-prover")]
+pub mod batch_prover;
 
 /// ERRORS
 /// ===============================================================================================